+++ app/src/main/java/com/talentmarketplace/viewmodel/JobPostingViewModel.kt
@@ -1,5 +1,6 @@
 package com.talentmarketplace.viewmodel
 
+import androidx.compose.runtime.MutableState
 import androidx.compose.runtime.mutableStateOf
 import androidx.lifecycle.ViewModel
 import dagger.hilt.android.lifecycle.HiltViewModel
